requirements.txt
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
openai>=0.27.2
blinker==1.5
blueprint

